env-info: route via AgentContext extended_content; remove inline append and legacy strip; show in Extended Context; fix config order ints; update tests
Fixes: oh-tab-1zt
