+++ .openhands/conversations/a9eb9407-53ed-45f4-b44d-e665cf6264e7/events.jsonl
@@ -1,0 +1,35 @@
+{"kind":"SystemPromptEvent","source":"agent","system_prompt":{"type":"text","text":"You are OpenHands, an autonomous AI agent running inside VS Code."},"tools":[{"type":"function","function":{"name":"terminal"}},{"type":"function","function":{"name":"file_editor"}},{"type":"function","function":{"name":"task_tracker"}},{"type":"function","function":{"name":"browser"}}],"id":"7ab4eaf6-9774-4bb4-910a-6a5b70b76f91","timestamp":"2025-11-19T06:24:56.006Z"}
+{"kind":"MessageEvent","source":"user","llm_message":{"role":"user","content":[{"type":"text","text":"hey"}]},"id":"249fcd0f-0644-463a-b989-65216bbfc0cf","timestamp":"2025-11-19T06:24:56.008Z"}
+{"agent_status":"RUNNING","kind":"ConversationStateUpdateEvent","source":"agent","id":"71270006-1da0-4132-b511-51686f27e75f","timestamp":"2025-11-19T06:24:56.010Z"}
+{"key":"llm_stream","value":"Hey","kind":"ConversationStateUpdateEvent","source":"agent","id":"0999e67f-8231-4a88-b394-94be4cb5be09","timestamp":"2025-11-19T06:24:58.626Z"}
+{"key":"llm_stream","value":"Hey!","kind":"ConversationStateUpdateEvent","source":"agent","id":"e91fb8fa-64e3-47a9-81e2-a5fbbd7863a9","timestamp":"2025-11-19T06:24:58.628Z"}
+{"key":"llm_stream","value":"Hey! How","kind":"ConversationStateUpdateEvent","source":"agent","id":"135250bf-3498-48a4-998f-855b426ff2c1","timestamp":"2025-11-19T06:24:58.665Z"}
+{"key":"llm_stream","value":"Hey! How can","kind":"ConversationStateUpdateEvent","source":"agent","id":"e92cc787-c30d-4bdc-8cdb-112a9fd0a574","timestamp":"2025-11-19T06:24:58.666Z"}
+{"key":"llm_stream","value":"Hey! How can I","kind":"ConversationStateUpdateEvent","source":"agent","id":"035e2954-f0dd-4556-b4cc-b168210beaeb","timestamp":"2025-11-19T06:24:58.707Z"}
+{"key":"llm_stream","value":"Hey! How can I help","kind":"ConversationStateUpdateEvent","source":"agent","id":"dce916ab-f5ca-4d89-a7e8-177e664279d6","timestamp":"2025-11-19T06:24:58.707Z"}
+{"key":"llm_stream","value":"Hey! How can I help you","kind":"ConversationStateUpdateEvent","source":"agent","id":"3d2ee1dd-f3cf-4f25-934f-318c511a501a","timestamp":"2025-11-19T06:24:58.743Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today","kind":"ConversationStateUpdateEvent","source":"agent","id":"fe80d5c6-47d4-4b53-a54b-3cf9f7ada343","timestamp":"2025-11-19T06:24:58.743Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?","kind":"ConversationStateUpdateEvent","source":"agent","id":"1e02f63f-7017-476c-be69-841c3c69f125","timestamp":"2025-11-19T06:24:58.868Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \n","kind":"ConversationStateUpdateEvent","source":"agent","id":"7a802484-ed27-4672-ad88-5d5f441592d4","timestamp":"2025-11-19T06:24:58.869Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed","kind":"ConversationStateUpdateEvent","source":"agent","id":"6767fba1-4b1f-4674-ae64-87a0d9fd0c14","timestamp":"2025-11-19T06:24:59.061Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a","kind":"ConversationStateUpdateEvent","source":"agent","id":"ab45f9c3-8ef0-4833-934f-47d2129affd7","timestamp":"2025-11-19T06:24:59.062Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand","kind":"ConversationStateUpdateEvent","source":"agent","id":"e52910b1-46ac-40bd-8422-014652678c95","timestamp":"2025-11-19T06:24:59.226Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with","kind":"ConversationStateUpdateEvent","source":"agent","id":"efb761c3-863f-4b03-9a9b-4be07dae4d5e","timestamp":"2025-11-19T06:24:59.228Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code","kind":"ConversationStateUpdateEvent","source":"agent","id":"cb66043f-1736-469a-86c4-9a06d24530a6","timestamp":"2025-11-19T06:24:59.252Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code,","kind":"ConversationStateUpdateEvent","source":"agent","id":"abc81643-27d1-4d0c-9fc9-d875fbddbac0","timestamp":"2025-11-19T06:24:59.253Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging","kind":"ConversationStateUpdateEvent","source":"agent","id":"a869316f-c35e-410f-8859-8f1194c931d3","timestamp":"2025-11-19T06:24:59.253Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging,","kind":"ConversationStateUpdateEvent","source":"agent","id":"0ca6684c-6c92-4872-81fb-c855915267e0","timestamp":"2025-11-19T06:24:59.254Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing","kind":"ConversationStateUpdateEvent","source":"agent","id":"aed70c3a-4515-4839-9fae-ac03e41ab74a","timestamp":"2025-11-19T06:24:59.275Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing,","kind":"ConversationStateUpdateEvent","source":"agent","id":"e68fd0f9-7412-4dde-8858-c18fb6ec807c","timestamp":"2025-11-19T06:24:59.275Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing, or","kind":"ConversationStateUpdateEvent","source":"agent","id":"3b564fd5-00e5-47a5-b029-f9931801e4f2","timestamp":"2025-11-19T06:24:59.319Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing, or something","kind":"ConversationStateUpdateEvent","source":"agent","id":"037894b8-488e-40e0-b3c2-24a8e11eb791","timestamp":"2025-11-19T06:24:59.319Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing, or something else","kind":"ConversationStateUpdateEvent","source":"agent","id":"e3c41840-eb8e-495e-86fe-7421ac89ab61","timestamp":"2025-11-19T06:24:59.376Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing, or something else in","kind":"ConversationStateUpdateEvent","source":"agent","id":"d3e68a10-f391-4187-b8cd-c8b375a89b5e","timestamp":"2025-11-19T06:24:59.377Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing, or something else in your","kind":"ConversationStateUpdateEvent","source":"agent","id":"17335efb-b57e-4ae0-b2ea-cf8caa7c6bc8","timestamp":"2025-11-19T06:24:59.440Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing, or something else in your project","kind":"ConversationStateUpdateEvent","source":"agent","id":"9072f77e-c493-4417-8794-0527c7febe17","timestamp":"2025-11-19T06:24:59.441Z"}
+{"key":"llm_stream","value":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing, or something else in your project?","kind":"ConversationStateUpdateEvent","source":"agent","id":"58c825e5-0d14-4d35-b5a0-810e086a00ab","timestamp":"2025-11-19T06:24:59.473Z"}
+{"key":"llm_usage","value":{"input":25,"output":36,"cacheRead":0},"kind":"ConversationStateUpdateEvent","source":"agent","id":"ed0bd436-32a9-4918-a30a-ac3fccc00297","timestamp":"2025-11-19T06:24:59.489Z"}
+{"kind":"MessageEvent","source":"agent","llm_message":{"role":"assistant","content":[{"type":"text","text":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing, or something else in your project?"}]},"id":"eae4dd14-61e1-43d7-a560-391fe9547db2","timestamp":"2025-11-19T06:24:59.490Z"}
+{"key":"llm_usage","value":{"inputTokens":25,"outputTokens":36,"cacheReadTokens":0},"kind":"ConversationStateUpdateEvent","source":"agent","id":"4e6d79d6-d806-4b16-af2a-211d21918972","timestamp":"2025-11-19T06:24:59.490Z"}
+{"iteration":1,"kind":"ConversationStateUpdateEvent","source":"agent","id":"4af31f69-bde0-477f-83d9-9a40a1a717ac","timestamp":"2025-11-19T06:24:59.490Z"}
+{"agent_status":"IDLE","kind":"ConversationStateUpdateEvent","source":"agent","id":"04b39bd7-d69b-4cd9-98d3-5c3812e65924","timestamp":"2025-11-19T06:24:59.490Z"}
