+++ .openhands/conversations/a9eb9407-53ed-45f4-b44d-e665cf6264e7/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":1,"values":{"llm_stream":"Hey! How can I help you today?  \nNeed a hand with code, debugging, writing, or something else in your project?","llm_usage":{"inputTokens":25,"outputTokens":36,"cacheReadTokens":0}}}
