+++ .openhands/conversations/local-mi5mnnkg/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":5,"values":{"llm_stream":"I can read it—what’s the path to the README in your repo (e.g., README.md), and do I have access to the workspace? If not, please paste the contents here.\n\nAlso, what would you like me to do with it?\n- Quick summary\n- Extract action items/todos\n- Check for inconsistencies or missing setup steps\n- Something elseI don’t currently have terminal access in your workspace. If you can grant me a shell or share the output, I can proceed.\n\nOptions:\n- Let me run a command: enable terminal access and I’ll run: ls -la\n- You run it and paste the output: run ls -la at the repo root and paste the result here\n- Share the file tree: e.g., tree -L 2 (if tree is installed) or code --list-extensions, etc., if you want me to review setup files\n\nHow would you like to proceed?I can run ls, but I don’t currently have access to your workspace or a terminal in this session—there’s no file tree in my context.\n\nIf you’re using me inside VS Code/OpenHands, please grant workspace/terminal access, then I’ll run:\n- ls -la\n- optionally: tree -a -L 2 (if tree is installed)\n\nAlternatively, you can paste the output of ls -la here, or share the file list, and I’ll proceed (e.g., locate README.md and summarize it).I can’t actually execute shell commands in this session—no terminal/filesystem tools are available to me right now.\n\nIf you want me to run ls:\n- In VS Code/OpenHands, enable my terminal/filesystem access, then ask me again.\n- Or run this and paste the output:\n  - pwd\n  - ls -la\n  - optionally: tree -a -L 2 (if installed)\n\nOnce I see the file list, I can locate and read the README for you.Yes. Quick recap:\n- You asked me to read the README.\n- I asked for the path or workspace access.\n- You told me to run ls.\n- I explained I don’t have terminal/workspace access here and asked you to grant it or paste the output.\n- You asked me to try anyway; I reiterated I can’t execute commands without access.\n\nIf you enable workspace/terminal access, I’ll run ls -la and open the README. Or paste the README (or ls output) here and I’ll summarize it.","llm_usage":{"inputTokens":510,"outputTokens":181,"cacheReadTokens":0}}}
