+++ .openhands/conversations/local-mi5p89cd/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":1,"values":{"llm_stream":"Hey! How can I help you today? If you're working on code, I can help debug, write functions, explain errors, or navigate files.","llm_usage":{"inputTokens":25,"outputTokens":39,"cacheReadTokens":0}}}
