+++ .openhands/conversations/local-mi5p8g5a/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":4,"values":{"llm_stream":"Hey! How can I help with .github/workflows/e2e.yml?\n\nIf you want a review or suggestions, please share the file contents here. I can:\n- Explain what each job/step does\n- Spot issues (security, caching, concurrency, permissions)\n- Suggest improvements (pin actions, OIDC auth, matrix, retries, artifacts, timeouts)\n- Help debug failing steps/logicHey! What would you like to do with .github/workflows/e2e.yml?\n\n- I can review it for issues and improvements\n- Help debug a failing run\n- Explain what each step does\n- Propose a cleaner or faster setup\n\nIf you paste the file (or the parts you care about) and tell me your goal, I’ll jump in.I see media/screenshots/README.txt in the repo. Do you want me to open it and summarize or review/suggest updates?I don’t have the contents of media/screenshots/README.txt yet. Please paste it here, or let me know if you want me to propose an improved README template and best practices for maintaining screenshots (naming, format, DPI, diffs, and automation).","llm_usage":{"inputTokens":287,"outputTokens":383,"cacheReadTokens":0}}}
