+++ .openhands/conversations/local-mi5pgj8z/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":2,"values":{"llm_stream":"Hey! How can I help you today? I can assist with coding, debugging, exploring files, writing, data analysis, or running commands in your workspace.I see README.md selected. What would you like me to do with it?\n- Summarize or explain it\n- Review for clarity/structure and suggest improvements\n- Fix formatting/grammar\n- Add sections (install, usage, examples, badges, TOC, license, contributing)\n- Generate a polished README template based on your project\n\nIf you want me to review the actual content, please let me open it or paste it here.","llm_usage":{"inputTokens":87,"outputTokens":419,"cacheReadTokens":0}}}
