+++ .openhands/conversations/local-mi5rpht1/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":3,"values":{"llm_stream":"Hey! How can I help you today? If you’re working on something in VS Code or need help with coding, debugging, or setting up a project, just let me know what you’re trying to do.I can read media/screenshots/README.txt. Do you want me to:\n- display the full contents, or\n- summarize the key points?\n\nLet me know your preference and I’ll proceed.I can summarize it, but I don’t have the file contents in this chat yet. Please either:\n- allow me to open media/screenshots/README.txt, or\n- paste its contents here.\n\nOnce I have the text, I’ll give you a concise summary.","llm_usage":{"inputTokens":159,"outputTokens":512,"cacheReadTokens":0}}}
