+++ .openhands/conversations/local-mif1poij/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":2,"values":{"llm_stream":"Hey! How can I help you today?","llm_usage":{"inputTokens":4913,"outputTokens":27,"cacheReadTokens":4864},"llm_tool_call":"call_0fLiBpWogL1j1HRAAMYjsLZi"}}
