+++ .openhands/conversations/local-mif32lzu/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":7,"values":{"llm_stream":"Hey! How can I help you today?I read README.md and created FACTS.txt with three summarized facts.\n\nPath: /Users/enyst/repos/oh-tab/FACTS.txt\n\nContents:\n1) OpenHands-Tab is a VS Code extension that lets you interact with OpenHands AI agents directly in the IDE, featuring chat, action confirmation with security indicators, conversation history, and integrated terminal output.\n2) It can run agents locally inside VS Code or connect remotely to an agent-server, and the project is organized as an npm workspace with a root extension and a TypeScript SDK (@openhands/agent-sdk-ts).\n3) Quick start requires VS Code 1.104+ and Node.js 22+, with common commands like `npm run build`, `npm test`, and `npm run lint` for development.","llm_usage":{"inputTokens":6621,"outputTokens":160,"cacheReadTokens":6272},"llm_tool_call":"call_SFPtj4xndwa7H5eyl3yOzkML"}}
