+++ .openhands/conversations/local-mihirx54/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":10,"values":{"llm_stream":"Hey! How can I help you today?The first time Mira heard the city speak, it arrived one syllable at a time, like rain teasing meaning from the air. Neon signs whispered half-words as they flickered, alleyway screens coughed up letters that tumbled into sense, and the old clocktower inhaled commas between its chimes. She learned to wait. To taste the pause after a consonant, to hold out her palm for the next vowel, to watch sentences bead along the midnight like a silver necklace forming itself. The city never dumped a story all at once. It streamed it, patient and particulate, as if it believed meaning should be earned in breaths.\n\nOne evening, a storm wandered in with a backpack full of lightning and the city’s voice broke into stutters. Mira followed the trail of half-phrases—“A door… below the… don’t forget to—”—down to the ferry docks, where the tide kept time by tapping the pylons. The wind ferried fragments across the water, and she cupped them like fireflies. Every gap grew into a small suspense, every pause an invitation. By the time the last token arrived—“…lantern”—she understood: a key hidden in a paper lantern drifting toward the far shore. She ran, laughing at the audacity of a message that trusted her to meet it halfway.\n\nThat night, Mira lit her own lantern, sending a reply in kind. She wrote in slow sparks, letting each word float free before the next took flame. Somewhere out there, someone else would gather the letters as they came, and in the small distance between her glow and their gaze, the story would become theirs too. She realized the city hadn’t simply been speaking; it had been teaching her a different grammar of attention, one where the spaces were part of the sentence, and arrival—bit by bit—was its own kind of magic.I read the README and saved three key facts to FACTS.txt. Here’s what I captured:\n- OpenHands-Tab is a VS Code extension for interacting with OpenHands AI agents, featuring a chat interface with streaming event display.\n- It supports local mode (agent runs in VS Code) and remote mode (connects to an agent-server), including action confirmation with security risk indicators.\n- Setup requires VS Code 1.104.0+ and Node.js 22+; install via git clone, npm install, npm run build, then launch with F5 and run “OpenHands: Open Tab”.","llm_usage":{"inputTokens":8940,"outputTokens":126,"cacheReadTokens":8576},"llm_tool_call":"call_loCFlEbCgE5DrV8zRGMYKuHC"}}
