+++ .openhands/conversations/local-mihl15aw/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":20,"values":{"llm_stream":"Hey! How can I help you today?I read README.md and created a concise summary at FACTS.txt.\n\nPath:\n- /Users/enyst/repos/oh-tab/FACTS.txt\n\nContents include:\n- Purpose of the extension\n- Key features\n- Prerequisites\n- Installation steps\n- Development workflow\n- Configuration options\n- Documentation index\n- Architecture overview and SDK components\n- Common npm commands\n- License (MIT)\n\nWould you like me to adjust the level of detail or add links/examples?Yes. I see:\n- /Users/enyst/repos/oh-tab/FACTS.txt (created just now)\n- /Users/enyst/repos/oh-tab/media/types/agent-sdk.js (already in the repo)\n\nDo you want me to read media/types/agent-sdk.js and summarize it as well?","llm_usage":{"inputTokens":12957,"outputTokens":65,"cacheReadTokens":12672},"llm_tool_call":"call_pf8V3tsA3gfLdT3xHI0dIEwG"}}
