+++ .openhands/conversations/local-mihm2z0z/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":7,"values":{"llm_tool_call":"call_v7uY82sx2xFco6rp5Al7VNhM","llm_usage":{"inputTokens":6658,"outputTokens":250,"cacheReadTokens":6528},"llm_stream":"Here’s the gist of AGENTS.md:\n\n- Purpose: Essential info for working on the OpenHands-Tab VS Code extension and its Agent SDK.\n- Prerequisites: Node.js 22 LTS (npm ≥ 10) and VS Code ≥ 1.104.\n- Quick commands:\n  - Install: npm ci\n  - Build all (SDK + extension + webview): npm run build\n  - Test: npm test (with example targeted Vitest command)\n  - Lint: npm run lint\n  - Type check: npm run typecheck\n  - Dev watch: npm run watch\n  - Launch extension: press F5 or run code \"$(pwd)\" --extensionDevelopmentPath=\"$(pwd)\"\n- Tooling setup: Ensure Node 22 and npm are on PATH (e.g., via nvm or tarball).\n- Project structure overview:\n  - src/: VS Code extension (entry, tests, connection/session/settings/sidebar/webview React UI)\n  - packages/agent-sdk-ts/: TypeScript SDK (conversation API, context, runtime, LLM clients, tools, types)\n  - tests/e2e/: End-to-end tests (truncated in view)"}}
