+++ .openhands/conversations/local-mj3wjpnm/state.json
@@ -1,0 +1,1 @@
+{"status":"idle","iteration":0,"values":{}}
