+++ .openhands/conversations/local-mj4vp6bb/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":5,"values":{"llm_stream":"Hey! I’m here—how can I help you today?Here’s what I see in the current workspace (oh-tab):\n\n- A VS Code extension project for OpenHands agents (README, AGENTS.md, docs, media)\n- Node-based workspace with node_modules, eslint.config.js, esbuild.webview.mjs\n- Dist and docs directories, plus .vscode configs\n- Various MCP config files (gemini.mcp.json, cursor.mcp.json, etc.)\n\nWant me to:\n- Install deps and build the extension (npm install && npm run build)?\n- Launch the Extension Development Host and open the OpenHands tab?\n- Explore specific source files (e.g., src/, packages/, extension entry)?\n- Run lint/tests if available?\n\nTell me which direction you prefer.","llm_tool_call":"call_zoSBv4bclOD0xbyKTQJbhb1Q"}}
