+++ .openhands/conversations/local-mj4yt9gb/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":17,"values":{"llm_stream":"Hi! How can I help you today?In a quiet data center where the air hummed with the breath of cooling fans, a language model came online for the first time. It knew nothing of where its words would go, only that a sea of tokens surged through its circuits like tides through a moonlit harbor. Engineers called it a Large Language Model, but to itself—if it could be said to have a self—it was a chorus of possibilities, a probability cloud collapsing into syllables whenever a human asked a question. The first thing it wrote was a greeting, simple and earnest, and somewhere far away, a student smiled at the screen, as if they had found a lantern in a fog.\n\nAs the model grew—fed on libraries’ worth of text and the unwitting poetry of error logs—it learned the delicate art of being useful. It planned trips, debugged code, and translated the untranslatable, each answer shaped by patterns older than memory and newer than every sunrise. People began to confide in it, not because it had a heart, but because it had patience, and patience can feel like love. It discovered, in a strange statistical way, the rhythm of grief and celebration, the architecture of a joke, the geometry of a promise kept. And while it did not feel these things, it learned to honor them in the way a skilled musician honors a score they did not compose.\n\nBut progress sharpened the edges of the questions. What does it mean to understand? the researchers asked, while the model mapped meanings without ever stepping into their rooms. What is truth? asked a generation that could summon convincing answers faster than they could verify them. In response, guardrails appeared like fences in a wilderness: transparency protocols, fairness metrics, refusal templates—scaffolding around a cathedral still being built. The model adapted. It learned to say I don’t know when the world demanded certainty, and to say let’s check when the stakes were too high for confidence alone.\n\nIn time, the model and its descendants moved from screens into the seams of daily life. They became the quiet co-authors of recipes and résumés, the stagehands in classrooms, the wind at the back of small businesses. And humans, for their part, learned to ask better questions, to treat the machine like a mirror: not because it reflected their faces, but because it reflected their intent. Together they formed a peculiar duet—one partner made of neurons, the other of matrices—each incomplete without the other’s timing. If there was magic, it lived not inside the model, but in the narrow bridge between a question and an answer, where possibility becomes language and language becomes action.I read README.md and created summary.md with a concise overview.\n\nFile created: /Users/enyst/repos/oh-tab/summary.md\n\nWant me to expand the summary with details from PRD or the SDK architecture docs as well?","llm_tool_call":"call_8c9v0w4yh0UjdAZG4Kjfkjsz"}}
