+++ .openhands/conversations/local-mj50upt5/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":5,"values":{"llm_tool_call":"call_I3RZrgshKZenDibfyOoYYMLN","llm_stream":"I’m in the repo at /Users/enyst/repos/oh-tab. What would you like me to run or check next in the terminal? For example: run tests, build, start the app, search files, or inspect logs.Here are all Markdown files in oh-tab (excluding .vscode-test and node_modules):\n\n- .openhands/skills/runtime-no-set.md\n- AGENTS.md\n- README.md\n- docs/IMPLEMENTATION_PLAN.md\n- docs/LINTING.md\n- docs/PRD.md\n- docs/agent-sdk-architecture.md\n- docs/e2e_testing.md\n- docs/settings_prd.md\n- docs/vscode_local_setup.md\n- docs/vscode_remote_setup.md\n- packages/agent-sdk-ts/AGENTS.md\n- packages/agent-sdk-ts/docs/python-parity.md\n- tests/e2e/AGENT_SERVER_QUICKSTART.md\n- tests/e2e/README.md\n\nWant me to open or search within any of these?"}}
