+++ .openhands/conversations/local-mj5xwaq5/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":25,"values":{"llm_stream":"Hey! How can I help today?I read README.md and created a concise summary at summary.md.\n\nPath: /Users/enyst/repos/oh-tab/summary.md\n\nWould you like it shortened further or expanded to include screenshots/links?Yes—I can see docs/agent-sdk-architecture.md. Want me to summarize it or include any parts of it in summary.md as well?I’ve updated summary.md to incorporate docs/agent-sdk-architecture.md.\n\nUpdated file:\n- /Users/enyst/repos/oh-tab/summary.md\n\nWant me to add a short code snippet showing Conversation() local vs remote usage or event handling examples?Could you clarify what you mean by “local bd”? Do you mean:\n- Local mode in OpenHands-Tab (agent runs in VS Code)?\n- A local database (DB)?\n- Something else (e.g., Bangladesh, business development, etc.)?\n\nIf it’s local mode in this project, I can explain how the Conversation() factory selects LocalConversation, event handling, terminal integration, and workspace security.","llm_tool_call":"call_DNAQN8hwm5doE32yb3ywEVzz"}}
