+++ .openhands/conversations/local-mj5y70us/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":1,"values":{"llm_stream":"Do you mean one of these?\n\n- Local DB (local database): Running a database (SQLite, PostgreSQL, MySQL, etc.) on your machine for development/testing.\n- Local Big Data (local BD): Running Hadoop/Spark locally (single-node) for development.\n- Local Business Development/Business Directory: Strategies or listings focused on local markets.\n- Local by Flywheel (often called “Local”): A local WordPress development tool.\n- Bangladesh (BD): “Local BD” as in local context within Bangladesh.\n\nTell me which one you mean, and what you want to do with it, and I’ll tailor the answer."}}
