+++ .openhands/conversations/local-mj65pf0c/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":12,"values":{"llm_stream":"Hey! How can I help you today?Done. I removed the untracked summary.md in the current directory and left everything else untouched.I’m sorry — since summary.md was untracked and removed with rm, there’s no built-in way to “undo” it via Git.\n\nWhat I checked:\n- Git history for summary.md: no record found.\n- Searched the repo for any copies/temp files: none found.\n\nPossible recovery options (if applicable on your system):\n- Check editor autosave/backups (e.g., VS Code backups, .vscode settings, or temp files outside the repo).\n- Check OS backups (Time Machine/Windows File History) or cloud backups.\n- If it was generated from another source, I can help regenerate it if you tell me how it was produced or what it contained.\n\nIf you can share what was in summary.md (even roughly), I’ll recreate it for you.I don’t see any GitHub-related token environment variables set in this shell right now.\n\nSearched for common names:\n- GITHUB_TOKEN, GH_TOKEN, GH_PAT, GIT_TOKEN, PERSONAL_ACCESS_TOKEN, GITHUB_*TOKEN, GH_*TOKEN\n- Also scanned env var names containing github, gh, token, or pat\n\nNone found.\n\nIf you set it yesterday in a different shell/session or via a one-off export, it wouldn’t persist. Options:\n- If you saved it in a .env file or shell profile (e.g., ~/.zshrc, ~/.bash_profile), I can look there if you want.\n- If it’s a Git credential helper entry, we can check git config and OS keychain.\n- If it’s in a GitHub CLI (gh) auth, we can check gh auth status.\n\nTell me how you set it and I’ll locate it or reconfigure it.Got it. If you need anything else, just let me know!","llm_tool_call":"call_jrwdngd3IetP2pWxmqLedAFh"}}
