+++ .openhands/conversations/local-mj65uxfp/state.json
@@ -1,0 +1,1 @@
+{"status":"RUNNING","iteration":15,"values":{"llm_stream":"A few things can make a “normal” VS Code terminal stop accepting typing. Try these quick checks in order:\n\n1) Make sure the terminal isn’t read-only\n- Look for a small padlock icon in the terminal tab. If it’s locked, click it or run: Command Palette → Terminal: Toggle Read-Only Mode.\n- Some extension/task terminals are read-only; ensure you created a regular terminal: Terminal → New Terminal (or “Create New Terminal (With Profile)”).\n\n2) Confirm focus and a live shell\n- Click inside the terminal and check for a blinking caret.\n- Try Ctrl+` (macOS: Cmd+`) to focus the terminal.\n- If you see “[Process exited]” or “Press any key to close…”, that terminal is closed—create a new one.\n- Use the dropdown next to the “+” in the terminal panel to pick a profile (e.g., Bash, PowerShell, zsh) and try again.\n\n3) See if keypresses are being intercepted\n- Run: Command Palette → Developer: Toggle Keyboard Shortcuts Troubleshooting. Press a few letters in the terminal. If VS Code shows keybindings firing instead of “type” going to the terminal, your keys are being captured by VS Code.\n- In Settings, try toggling: Terminal › Integrated: Send Keybindings To Shell (on/off) and test both ways.\n\n4) Try paste\n- If right-click → Paste (or Ctrl/Cmd+V) works but typing doesn’t, it’s likely a keybinding/focus or extension issue.\n\n5) Disable extensions\n- Extensions can hijack input. Restart VS Code with extensions disabled:\n  - Command Palette → Reload Window With Extensions Disabled\n  - Or start VS Code with --disable-extensions and test.\n\n6) Reset/verify terminal profile\n- Settings → Terminal › Integrated: Default Profile. Pick a known-good shell (e.g., “Command Prompt” on Windows, “bash/zsh” on macOS/Linux).\n- Then Terminal → New Terminal.\n\n7) Basic resets\n- Close all terminals and open a new one.\n- File → Close Folder/Workspace, then open an empty window and open a new terminal there to rule out workspace settings.\n- Help → Toggle Developer Tools → Console to check for terminal-related errors.\n\nIf it’s still stuck, share:\n- OS and VS Code version\n- Does the terminal show a normal prompt?\n- Whether the padlock icon appears\n- Output from Keyboard Shortcuts Troubleshooting when you press keys\n- Whether the issue persists with extensions disabled and in a fresh empty window\n\nWith that, I can pinpoint the cause faster.","llm_tool_call":"call_TJX7EZmxbO2T47wJ2lKwb7Jt"}}
