+++ .openhands/conversations/local-mj6hf7js/state.json
@@ -1,0 +1,1 @@
+{"status":"IDLE","iteration":14,"values":{"llm_tool_call":"call_toUjjRrqEw02S0XduDRM50Nv","llm_stream":"No, there isn’t a summary.md in the current working directory.Created summary.md in the project root based on the README.\n\nHighlights included:\n- What OpenHands-Tab is and key features\n- Quick start (prereqs, install, build)\n- Development workflow\n- Configuration options\n- Documentation index with links\n\nWant me to tweak the format or add anything (e.g., architecture overview, screenshots, or commands for packaging/publishing)?# OpenHands-Tab Prototype\n\n<img width=\"1706\" height=\"952\" alt=\"image\" src=\"https://github.com/user-attachments/assets/9e430fe1-ba13-42d7-8c90-211cdf028abc\" />\n\nA VS Code extension for interacting with OpenHands AI agents directly in your IDE.\n\n## Features\n\n- Chat interface with streaming event display\n- Local mode (runs agent in VS Code) or remote mode (connects to agent-server)\n- Action confirmation with security risk indicators\n- Conversation history and persistence\n- Workspace file context and skills support\n- Integrated terminal output\n\n## Quick Start\n\n### Prerequisites\n\n- VS Code 1.104.0+\n- Node.js 22+\n\n### Installation\n\n```bash\ngit clone https://github.com/enyst/OpenHands-Tab.git\ncd OpenHands-Tab\nnpm install\nnpm run build\n```\n\n### Development\n\n1. Open the project in VS Code\n2. Press F5 to launch Extension Development Host\n3. Click the OpenHands icon in the Activity Bar (or run \"OpenHands: Open\") to reveal the chat sidebar view\n\n### Configuration\n\n- **OpenHands: Configure** - Set up server URL, LLM settings, API keysHey! How can I help you today? (Need anything in the repo — edits, more docs, run something, or answer a question?)"}}
