+++ packages/agent-sdk-ts/.openhands/TASKS.json
@@ -1,0 +1,7 @@
+[
+  {
+    "title": "Do work",
+    "notes": "",
+    "status": "todo"
+  }
+]
