+++ src/extension.ts
@@ -26,6 +26,7 @@ import { createPastedImagesCleanupScheduler } from './extension/pastedImagesClea
 import { registerSecretCommands } from './extension/secretCommands';
 import { summarizeWithLocalLlm } from './extension/summarizeWithLocalLlm';
 import { createHalConfigurationChangeHandler } from './extension/halConfigurationChangeHandler';
+import { formatEnvironmentInformation } from './shared/environmentInformation';
 import {
   createOutputLogger,
   normalizeOutputVerbosity,
@@ -157,6 +158,7 @@ function resolveActiveEditorFilePath(editor: vscode.TextEditor | undefined): str
 function syncActiveEditorSystemMessageSuffix(editor: vscode.TextEditor | undefined): void {
   activeEditorFilePath = resolveActiveEditorFilePath(editor);
   if (!localAgentContext) return;
+  // Only populate systemMessageSuffix; environment context for user messages will be provided via userMessageSuffix when sending.
   localAgentContext.systemMessageSuffix = activeEditorFilePath
     ? `Currently opened in the editor: ${activeEditorFilePath}`
     : undefined;
@@ -166,9 +168,24 @@ function syncActiveEditorSystemMessageSuffix(editor: vscode.TextEditor | undefin
  * Initialize the OpenHands extension: create logging channel and chat webview, register commands and configuration handlers, and wire up terminal, conversation, and secret-management behavior.
  *
  * Performs extension startup work and registers disposables (commands, webview provider, event listeners, and configuration change handlers) on the provided VS Code extension context.
- *
- * @param context - The VS Code extension context used to register disposables, access workspace and global state, and resolve extension resources
  */
+
+function buildEnvironmentInfoSuffix(): string | null {
+  try {
+    const workspaceRoot = vscode.workspace.workspaceFolders?.[0]?.uri.fsPath;
+    const activeEditorPath =
+      vscode.window.activeTextEditor?.document?.uri?.scheme === 'file'
+        ? vscode.window.activeTextEditor.document.uri.fsPath
+        : null;
+    const openEditorPaths = (vscode.window.visibleTextEditors ?? [])
+      .map((e) => (e?.document?.uri?.scheme === 'file' ? e.document.uri.fsPath : null))
+      .filter((p): p is string => typeof p === 'string' && p.length > 0);
+    return formatEnvironmentInformation({ workspaceRoot, activeEditorPath, openEditorPaths });
+  } catch {
+    return null;
+  }
+}
+
 export function activate(context: vscode.ExtensionContext) {
   const secrets = new SecretRegistry(context.secrets);
   secretRegistry = secrets;
@@ -368,8 +385,8 @@ export function activate(context: vscode.ExtensionContext) {
           markPrintedExitFor(cid);
         }
       }
-    } catch (e) {
-      console.error('[Terminal] Failed to write terminal event:', e);
+    } catch (err) {
+      console.error(`[Terminal] Failed to write terminal event: ${renderError(err)}`);
     }
   };
 
@@ -449,6 +466,13 @@ export function activate(context: vscode.ExtensionContext) {
     const needsNewConversation = !conversation || conversationMode !== desiredMode;
 
     if (needsNewConversation) {
+    // Populate AgentContext userMessageSuffix for environment info in local mode
+    if (desiredMode === 'local') {
+      if (!localAgentContext) localAgentContext = new AgentContext({ loadUserSkills: true });
+      const env = buildEnvironmentInfoSuffix();
+      localAgentContext.userMessageSuffix = env ?? undefined;
+    }
+
       try {
         conversation?.removeAllListeners();
         conversation?.disconnect();
@@ -457,8 +481,8 @@ export function activate(context: vscode.ExtensionContext) {
 
       const persistenceDir =
         desiredMode === 'local'
-          ? await resolveConversationStoreRoot({ context, getOutputChannel: () => outputChannel, renderError }).catch((err: unknown) => {
-              outputChannel?.appendLine(`[storage] Failed to resolve conversation store root: ${renderError(err)}`);
+          ? await resolveConversationStoreRoot({ context, getOutputChannel: () => outputChannel, renderError }).catch(() => {
+              outputChannel?.appendLine(`[storage] Failed to resolve conversation store root`);
               return path.join(os.tmpdir(), 'openhands-conversations-vscode');
             })
           : undefined;
