+++ src/webview-src/components/eventBlocks/MessageEventBlock.tsx
@@ -10,11 +10,6 @@ import {
   withAlpha,
 } from './shared';
 
-// Strip environment info blocks from display, but keep them in the original message sent to the LLM.
-const ENV_BLOCK_RE = /<environment information>[\s\S]*?<\/environment information>/gi;
-function stripEnvironmentInfo(text: string): string {
-  return text.replace(ENV_BLOCK_RE, '').trim();
-}
 
 /**
  * Renders chat messages - user and agent messages with context files,
@@ -26,7 +21,6 @@ export function MessageEventBlock({ event, index }: { event: AgentMessageEvent; 
   const isAgent = message.role === 'assistant';
 
   const rawText = message.content.filter(isTextContent).map((c) => c.text).join('\n');
-  const displayText = stripEnvironmentInfo(rawText);
   const CONTEXT_HEADER = 'User has selected the following files for you to read:';
   function parseContextBlock(text: string): { main: string; files: string[] } {
     const idx = text.lastIndexOf(CONTEXT_HEADER);
@@ -37,7 +31,7 @@ export function MessageEventBlock({ event, index }: { event: AgentMessageEvent; 
     const files = after.split(/\r?\n/).map((l) => l.trim()).filter((l) => l.length > 0);
     return { main: before, files };
   }
-  const { main: withoutContext, files: contextFiles } = parseContextBlock(displayText);
+  const { main: withoutContext, files: contextFiles } = parseContextBlock(rawText);
 
   const ATTACHMENT_BEGIN_LINE_RE = /^-{5,}\s*BEGIN ATTACHMENT:\s*(.*?)\s*-{5,}\s*$/;
   const ATTACHMENT_END_LINE_RE = /^-{5,}\s*END ATTACHMENT:\s*(.*?)\s*-{5,}\s*$/;
