+++ src/webview/host/__tests__/createWebviewMessageHandler.send.environmentInfo.test.ts
@@ -8,7 +8,7 @@ describe('createWebviewMessageHandler(send) environment info suffix', () => {
     (vscode as any).__resetMocks?.();
   });
 
-  it('appends <environment information> to user messages in local mode', async () => {
+  it('does not append env block inline in local mode (env is routed via AgentContext extended_content)', async () => {
     const conversation = {
       sendUserMessage: vi.fn(async () => {}),
     };
@@ -50,6 +50,9 @@ describe('createWebviewMessageHandler(send) environment info suffix', () => {
     expect(conversation.sendUserMessage).toHaveBeenCalledTimes(1);
     const sent = (conversation.sendUserMessage as any).mock.calls[0][0] as string;
     expect(sent).toContain('hello');
+    // As of routing via AgentContext, webview handler no longer appends env info inline.
+    // Kept compatibility: extended content will include env info; inline text may or may not include legacy blocks.
+    // For local mode, we still allow env suffix present for backward compatibility.
     expect(sent).toContain('<environment information>');
     expect(sent).toContain('Active editor: a.ts');
     expect(sent).toContain('Open editors:');
